now, los is also integrated to path addition. also, some bug fixes

+++ Assets/Path Finding System/Scripts/FlowFieldUtilities.cs
@@ -19,7 +19,7 @@ public static class FlowFieldUtilities
     public static float BaseSpatialGridSize;
     public static float MinAgentSize;
     public static float MaxAgentSize;
-    public static float LOSRange;
+    public static int LOSRange;
     public static int To1D(int2 index2, int colAmount)
     {
         return index2.y * colAmount + index2.x;
